fix(idv): remove redundant type cast in for-loop increment -- src/id/idv/idv.cpp
Note:
[ fix warnings ]

src/id/idv/idv.cpp:1198:10: warning: expression result unused [-Wunused-value]
 1198 |          (UInt)i ++)
      |          ^     ~~~~
src/id/idv/idv.cpp:1267:17: warning: expression result unused [-Wunused-value]
 1267 |                 (UInt)i ++)
      |                 ^     ~~~~

+++ src/id/idv/idv.cpp
@@ -1195,7 +1195,7 @@ void   idvManager::applyStatisticsToSystem(idvSession *aCurr, idvSession *aOld)
 
     for (i = IDV_STAT_INDEX_BEGIN;
          i < IDV_STAT_INDEX_MAX;
-         (UInt)i ++)
+         i++)
     {
         if ( (mStatName[i].mAttr  & IDV_ATTR_ACCUM) != 0)
         {
@@ -1264,7 +1264,7 @@ void idvManager::applyOpTimeToSession(
     {
         for (i = IDV_STAT_INDEX_BEGIN_STMT_TO_SESS;
                 i < IDV_STAT_INDEX_MAX;
-                (UInt)i ++)
+                i++)
         {
             if ( (mStatName[i].mAttr & IDV_ATTR_ACCUM) != 0)
             {
